// PokemonApp/src/app/api/models/pokemon.ts
/* tslint:disable */
/* eslint-disable */
export interface Pokemon {
  attacks?: {  };
  classification?: string;
  evolutionRequirements?: {  };
  evolutions?: Array<{  }>;
  fleeRate?: number;
  height?: {  };
  id: string;
  favourite: boolean;
  maxCP?: number;
  maxHP?: number;
  name?: string;
  resistant?: Array<string>;
  types?: Array<string>;
  weaknesses?: Array<string>;
  weight?: {  };

  [key: string]: any;
}